Updates performance and styles relating to locales/chat ui
