+++ app/[locale]/page.tsx
@@ -1,6 +1,4 @@
 import { Chatbot } from "@/components/Chatbot";
-import { BlackHole } from "@/components/canvas/BlackHole";
-import { Common, View } from "@/components/canvas/View";
 import About from "@/components/sections/About";
 import Contact from "@/components/sections/Contact";
 import Endnote from "@/components/sections/Endnote";
@@ -9,10 +7,6 @@ import Work from "@/components/sections/Work";
 export default function Page() {
   return (
     <main>
-      <View orbit className="black-hole">
-        <BlackHole />
-        <Common />
-      </View>
       <About />
       <Work />
       <Contact />
